cambiamenti per session id utente

+++ konga/src/it/konga/framework/kObjects/KAbstract_Dto.java
@@ -5,7 +5,7 @@ import java.io.ObjectInputValidation;
 import java.io.Serializable;
 
 /**
- * classe base per i DTO.<br>
+ * Classe base per i DTO.<br>
  * Utile come gerarchica base
  * @author Giampaolo Saporito
  * @date 24-09-2014
@@ -13,12 +13,12 @@ import java.io.Serializable;
 public abstract class KAbstract_Dto implements Serializable, ObjectInputValidation
 {
 	private static final long serialVersionUID = 7163853062626L;
-	private Integer id;
+	private Long id;
 
-	public Integer getId(){
+	public Long getId(){
 		return id;
 	}
-	public void setId(Integer id){
+	public void setId(Long id){
 		this.id = id;
 	}
 
@@ -28,4 +28,5 @@ public abstract class KAbstract_Dto implements Serializable, ObjectInputValidati
 		//validare l'oggetto
 		if( getId() < 0) throw new InvalidObjectException("L'id deve essere positivo");	
 	}
+
 }
